Added "getStaticProps" & "getStaticPaths" functions
For displaying details about the meetups with a dynamic id. Both function are required, "getStaticPaths" returns an object with two properties. "fallback" has been set to "false" so that during the build process all pages for different ids are built. "getStaticProps" function takes "context" as a parameters and inside the function this give access to an object with various properties. Finally, this function also returns and object for meet up data.